backend/requirements.txt
# Django & Core
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-redis==5.4.0
django-environ==0.11.2

# Database
psycopg2-binary==2.9.10

# API & Auth
djangorestframework-simplejwt==5.4.0
drf-spectacular==0.28.0

# Task Queue
celery[redis]==5.4.0
redis==5.2.1

# LLM & AI
langchain==0.3.14
langchain-openai==0.3.0
langgraph==0.2.60
openai==1.58.1

# Research Sources
arxiv==2.1.3

# Utilities
tenacity==9.0.0
python-dotenv==1.0.1
gunicorn==23.0.0
httpx==0.28.1
aiohttp==3.11.11
pydantic==2.10.4

# MCP Protocol
mcp==1.2.0

# Export & Reporting
markdown==3.7
weasyprint==63.1
openpyxl==3.1.5
